requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
opencv-python-headless==4.11.0.86
numpy==2.2.3
Pillow==11.1.0
ezdxf==1.3.5
reportlab==4.4.3
psycopg[binary]==3.2.9
